refactor: shop_v2 제거, shop 페이지에 합치기
