+++ apps/web/src/app/[locale]/shop/PetGotcha/PetGotcha.tsx
@@ -21,7 +21,10 @@ export function PetGotcha() {
         <Button size="l" onClick={() => setOpenModal('one-pet')}>
           1 Pet / 1,000 P
         </Button>
-        <Button size="l">10 Pets / 10,000 P</Button>
+        <Button size="l" disabled>
+          Coming Soon!
+        </Button>
+        {/* <Button size="l">10 Pets / 10,000 P</Button> */}
       </div>
       <div className={petContainerStyle}>
         <Image src="/shop/pet-gotcha-image-card.webp" alt="pet-gotcha-pet" width={1120} height={530} />
