--- apps/web/src/app/[locale]/shop_v2/page.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-import { PetGotcha } from './PetGotcha';
-
-function ShopPage() {
-  return (
-    <div>
-      <PetGotcha />
-    </div>
-  );
-}
-
-export default ShopPage;
